Add build and runner scripts with Docker support
Added bash scripts 'build' and 'runner' to aid in application building and running. Implemented Docker support to containerize the application, included a Dockerfile, and used a 'Makefile' to simplify build and deployment commands. Additionally, added a '.gitignore' file to exclude unnecessary files from the repository and created a 'docker-compose.yml' file to support docker-compose up commands for easy deployment. These changes aim to streamline and automate the build and deployment process.